Make one function that displays a WR menu, this allows me to work on custom styles in the near future!
Also, show your rank in the map (if it's top50, that's the best I could
pull with just one query and without caching too much stuff) and number
of records.
TODO: Remove sm_wrsw and make sm_wr one menu with all difficulties.
TODO: Add style configuration through mysql (`style` table) and allow it
to be custom with many options.
TODO: Add a cvar that limits the amount of records in the WR menu.