Added dynamic zone settings.
If you edit the plugin to add more zone tracks, now their zone settings will be respected too. To use this feature, copy the zone settings from the first one to the last one of each track, rename and modify. i.e. the diff for shavit-zones.cfg in this commit.