Implemented SetStart/DeleteSetStart into zones
The player's position and view angles when using sm_r, sm_b, and sm_start are saved into the startpositions table in the database. The primary key for each entry is a concatenation of their Steam ID, map name, and track number. It still needs to be tested on MySQL. The Natives have been tested in a separate plugin.