boomshot stages stuff
Move remaining shavit-rankings dependent calls (#1043)

If the shavit-rankings plugin isn't loaded, the calls to its natives will cause an exception and the client's messages will not be formatted correctly.

Add native for getting the number of stages

Implemented native for getting number of stages and allowed practice mode to view stage times.

Added WR stage time comparisons for when a user enters a stage zone.

It's important that the stage zone number starts at two. It may be preferred to remove the user input of stage numbers and just auto increment as stages are added. Stage 1 should be the start zone and just have any additional stages increment from there. I'll readdress this if it's necessary.

Add translations for WR stage time comparison messages

Accidentally undid a change from a previous commit.

I didn't have a merge conflict, so I'm unsure why it got replaced.

Combine stage time translations into one

Implemented suggestions to translations and changed the stagetimes table

The stagetimes table no longer uses auth as a primary key, so it will now replace the old WR stage time.
However, the auth is still stored, so it can be used to reference the player and potentially a players PB.

Update callback name for replacing stage time for logging.