shavit-misc Add cvar UnNoclipOnRestart (#1232)
* shavit-misc Add cvar UnNoclipOnRestart


Add functionality to turn off noclip on timer restart, disabled by default u.u

I used to use a separate plugin I made for this, but figured it should be in the main plugin itself as a cvar instead. Since adding as a cvar, may as well see if it is wanted upstream.

* shavit-misc.sp shavit_misc_unnocliponrestart 1 by default (^: