shavit-checkpoints.sp - fix menu persistence
the menu now reappears after a menu that interrupted it is closed. i believe this was supposed to be intended behavior, but wasn't handled properly.

menu persistence now works very well, and the checkpoint menu will reappear unless you manually close it with "0". if you change to another kzcheckpoint or segmented style, it will again reappear.

only minor bug is that if you change between kzcheckpoint <-> segmented styles, the menu must be redrawn with the new options, so it "interrupts" itself, making you have to press 0 twice to close it in that instance.. so minor i bet nobody would even notice.. but i did.. so i mentioned it..